Lab 3

+++ FinanceManager.DAL/Entities/Account.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 
 namespace FinanceManager.DAL.Entities
 {
@@ -8,7 +7,7 @@ namespace FinanceManager.DAL.Entities
         public int Id { get; set; }
         public string Number { get; set; }
         
-        public int Count { get; set; }
+        public decimal Count { get; set; }
         
         public ICollection<Transaction> TransactionsAsSource { get; set; }
